Fix some issues
Image velocity is now based on the width of the screen so that it always moves in a similar way regardless of size.
Fixed flashing when on window resize.
Fixed problem where canvas would not cover the whole viewport.
Position is now maintained while resizing unless the image is outside of the canvas.